Move createRuntimeState inside
